requirements.txt
pynput
pyyaml
psutil
python-mpv 